Add lecture creation functionality to course API
Implemented a feature in the course API to allow for the creation of lectures associated with courses. Introduced a new LectureCreateSchema for validating the input data for lectures. Modified the CourseService to include an add_lecture method that appends a new lecture to the course using an atomic push operation to ensure no data loss or overwrite occurs during concurrent updates.